Update GibbsFunction_Kostas_draft.py
Jeff, regarding the last changes, I would like to ask you tomorrow, but to me, there is a conflict on the way we calculate the Ae (Helmholtz in excess). 
Also i realised that the whole calculation needs to be done 2 time, one for the vapour and one for the liquid phase, correct? 

I will meet you tomorrow before my tutorial at 12 or after 1 to explain you what the code reads at the end. 

Thank you, 

Kostas